apis/stores/pyproject.toml
[project]
name = "que-vino-stores"
version = "0.1.0"
description = "Que Vino Stores API"
authors = [
    {name = "Gustavo"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "google-cloud-bigquery (>=3.40.1,<4.0.0)",
    "firebase-admin (>=7.3.0,<8.0.0)",
    "mcp (>=1.26.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
